(request.js) EventEmitter memory leak
Hi,

When crawling a link for robots.txt with a redirect loop,  request.js throws the following error:

`(node:23420) Warning: Possible EventEmitter memory leak detected. 11 pipe listeners added. Use emitter.setMaxListeners() to increase limit`

It will be good to have a limit on the maximum Redirects. Google supports upto 4 redirects for robots.txt 

I have update robotsUrl to an object of options with maxRedirects set to 4 (defaults to 20). It will prevent crashing of the process and will also not waste server resource.